requirements.txt

Flask==3.1.0
folium==0.14.0
shapely==2.0.1
matplotlib==3.7.1
pandas==2.2.3
numpy==1.24.3
ipykernel==6.19.2
jupyter==1.0.0
openmeteo-py==0.0.1
openmeteo-requests==1.3.0
requests_cache
retry_requests
urllib3
cython
statsmodels==0.14.4
pycosmos~=0.11
requests-cache~=1.2.1
pycosmo~=2.2.0
fonttools~=4.56.0
seaborn
tqdm==4.67.1
rpy2==3.5.17
flask-cors==5.0.0
flask-restful==0.3.9
SALib==1.5.1
flask-swagger==0.2.14
flasgger==0.9.7.1